TODO item for later. Doxygen fixes and lint fix.
git-svn-id: file:///svn/unbound/trunk@147 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/mini_event.c
@@ -84,7 +84,8 @@ void *event_init(void)
 	if((int)FD_SETSIZE < base->capfd)
 		base->capfd = (int)FD_SETSIZE;
 #endif
-	base->fds = (struct event**)calloc(base->capfd, sizeof(struct event*));
+	base->fds = (struct event**)calloc((size_t)base->capfd, 
+		sizeof(struct event*));
 	if(!base->fds) {
 		event_base_free(base);
 		return NULL;
